requirements.txt
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
python-telegram-bot==21.6
PyYAML==6.0.2
sniffio==1.3.1
